chore(engine): day 2 — move types only to @node9/policy-engine
Per the v1.16.1 migration plan, day 2 moves type definitions out of
node9-proxy and into the engine package. Zero runtime change — types
erase at compile time. Implementations stay where they are; they
move on day 3+.

Types moved to packages/policy-engine/src/types.ts:
- SmartCondition, SmartRule (from src/config/index.ts)
- DlpMatch (from src/dlp.ts)
- RiskMetadata (from src/context-sniper.ts)

The engine's index.ts re-exports them. The proxy's existing files
(src/dlp.ts, src/config/index.ts, src/context-sniper.ts) re-export
from the package, so every existing import path in the codebase still
works unchanged. The package is now the source of truth; the proxy
files are thin re-exports.

Two TS-narrowing fixes were needed:
- src/__tests__/shields.test.ts: explicit param type on .every()
- src/mcp-server/index.ts: explicit param type on .map()

Both because Zod's inferred condition type and the package's
SmartCondition are structurally compatible but TypeScript can't
narrow across the package boundary without help. These are typing
hints, not behavior changes.

Verification:
- All 1378 tests pass (22 skipped, no regressions).
- npm run typecheck clean across both proxy and engine package.
- npm run lint: 0 errors, 8 pre-existing warnings unchanged.
- npm run format: clean.
- Engine package builds successfully.

Plan reference: doc/roadmap/v1.16.1-policy-engine.md
