+++ packages/policy-engine/src/index.ts
@@ -5,8 +5,9 @@
 // definitions live in exactly one place. No fs/process/os imports at
 // runtime — host code passes paths and state in.
 //
-// Day 1: package skeleton only. The first real export comes on Day 2
-// (types) and Day 3 (DLP module).
+// Day 2: types only. Implementations stay in node9-proxy until Day 3+.
 
 /** Engine version stamped on audit entries for future drift detection. */
 export const ENGINE_VERSION = '0.1.0-alpha.0';
+
+export type { SmartCondition, SmartRule, DlpMatch, RiskMetadata } from './types';
